Rename index files and fix broken story
