+++ src/Calendar/components/TimeGrid/components/Event/Event.scss
