--- src/Calendar/index.js
@@ -1,4 +1,0 @@
-import DayGrid from './components/DayGrid';
-import TimeGrid from './components/TimeGrid';
-
-export { DayGrid, TimeGrid };
