+++ src/EventGroupSelect/EventGroupSelect.js
