--- src/components/CalendarMonth/utils/index.js
@@ -1,3 +1,0 @@
-import getMonthGrid from './getMonthGrid';
-
-export { getMonthGrid };
