+++ src/index.js
@@ -1,12 +1,12 @@
-import FullCalendar from './components/FullCalendar';
-import CalendarGroups from './components/CalendarGroups';
-import CalendarMonth from './components/CalendarMonth';
-import CalendarMonthHeatmap from './components/CalendarMonthHeatmap';
-import CalendarWeek from './components/CalendarWeek';
-import DayList from './components/DayList';
-import DayGrid from './Calendar/components/DayGrid';
+import FullCalendar from './components/FullCalendar/FullCalendar';
+import CalendarGroups from './components/CalendarGroups/CalendarGroups';
+import CalendarMonth from './components/CalendarMonth/CalendarMonth';
+import CalendarMonthHeatmap from './components/CalendarMonthHeatmap/CalendarMonthHeatmap';
+import CalendarWeek from './components/CalendarWeek/CalendarWeek';
+import DayList from './components/DayList/DayList';
+import DayGrid from './Calendar/components/DayGrid/DayGrid';
 import TimeGrid from './Calendar/components/TimeGrid/TimeGridWrapper';
-import Toolbar from './components/Toolbar';
+import Toolbar from './components/Toolbar/Toolbar';
 
 export default FullCalendar;
